backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pyodbc==5.2.0
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3

